solve circular dependency
